Fixing remaining projects with netsandard configurations that won't build succesfully
