+++ src/Microsoft.XmlSerializer.Generator/tests/Configurations.props
@@ -2,7 +2,6 @@
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
     <BuildConfigurations>
-      netstandard;
       netfx;
       uap;
       netcoreapp;
